--HG--
branch : fw4spl_0.6

+++ SrcLib/core/fwData/test/tu/src/GenericFieldTest.cpp
@@ -27,7 +27,7 @@ void GenericFieldTest::tearDown()
 
 void GenericFieldTest::methode1()
 {
-    typedef double TestType;
+    typedef float TestType;
 
     const TestType VALUE = 13.456789 ;
